Add script import/reload toggle
Adds a Project Setting to toggle the automatic importing and reloading of LuaScripts, as well as allowing for the first-line comments `--!noimport` (to disable importing for a specific script, when the global toggle is `true`) and `--!doimport` (to enable importing for a specific script, when the global toggle is `false`). This comes in handy when one is editing a Lua script in the Godot editor that is not intended to be attached to a node (for example, a library), and the editor's automatic reloading and execution of the script for every change is undesired.
Additionally, move project settings out of `LuaScriptLanguage.cpp` and into its own file, for other classes' usage.